+++ CHANGELOG.md
@@ -98,6 +98,7 @@
 - Add WAP value in StockLocation report
 - Contract: Removed wrong demo data and added demo data for app configuration.
 - Analytic move line : accountType added on grid view.
+- Inventory:Allow to do an inventory for one product.
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.